+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 | ---------------- | ------- | ------------------------------------------------------------------ |
 | Symbol Bootstrap | v1.0.4  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
 
+-   Added `wizard` command.
+-   Added `pack` command.
 -   Fixed case issue validating keys when creating certificates.
 
 ## [1.0.3] - Mar-31-2021
@@ -22,8 +24,6 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 | ---------------- | ------- | ------------------------------------------------------------------ |
 | Symbol Bootstrap | v1.0.3  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
 
--   Added `wizard` command.
--   Added `pack` command.
 -   Improved Custom Preset Object types for symbol bootstrap lib integration.
 -   TransactionSelectionStrategy's new default value is `oldest`.